# specs/server-api.openapi.yaml
...
openapi: 3.1.0
info:
  title: Barca current HTTP API
  version: '1'
  description: 'Normative current wire contract; specification version 1 is independent
    of package versions. Scope: registered barca-server routes and fallback responses.
    No authentication is implemented (security is empty); read-only is an operating
    mode. No proposed RFC command routes, authorization, freshness eligibility or
    partition retrieval API are advertised. Conformance: crates/barca-server/tests/openapi.rs
    and python/tests/test_http_openapi_client.py.'
...
servers:
- url: http://127.0.0.1:8274
  description: Default local listener; configurable host/port, reverse proxies may
    add a prefix.
security: []
paths:
...
  /run/{target}:
    post:
      summary: Run one task and refresh its upstream assets
      operationId: post_run_target
      responses:
        '200':
          description: Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunHandle'
        '400':
          description: Invalid source/target or wrong node kind
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
            text/plain:
              schema: *id002
        '403':
          description: Read-only server refuses mutation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Ambiguous target or terminal run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Execution/inspection infrastructure error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '405':
          description: Registered path with unsupported method; Allow header identifies
            supported methods
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Allow:
              schema: *id002
              description: Methods accepted by this route
      description: Bodyless trigger. Target may be a bare function name or full node
        ID. This catch-all path accepts embedded slashes (encoded or unencoded).
      parameters:
      - name: target
        in: path
        required: true
        schema: *id002
    delete:
      summary: Cancel an in-flight run by polling handle
      operationId: delete_run_target
      responses:
        '200':
          description: Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cancellation'
        '403':
          description: Read-only server refuses mutation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Ambiguous target or terminal run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '405':
          description: Registered path with unsupported method; Allow header identifies
            supported methods
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Allow:
              schema: *id002
              description: Methods accepted by this route
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
            text/plain:
              schema: *id002
      description: Here target is the polling handle, not a node name. Returns cancelling
        immediately; poll /status for cancelled. A terminal run returns 409; a durable
        ID without a retained handle cannot be cancelled.
      parameters:
      - name: target
        in: path
        required: true
        schema: *id002
...
  /runs/{id}:
    get:
      summary: Inspect durable ID or retained live handle
      operationId: get_runs_id
      responses:
        '200':
          description: Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunDetail'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Execution/inspection infrastructure error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '405':
          description: Registered path with unsupported method; Allow header identifies
            supported methods
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers: &id015
            Allow:
              schema: *id002
              description: Methods accepted by this route
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
            text/plain:
              schema: *id002
      description: Durable steps and logs survive restart. Engine result/final output
        and cached per-step reports are available only while the live handle is retained,
        then result is null. Inspection uses a private DB snapshot.
      parameters: &id016
      - name: id
        in: path
        required: true
        schema: *id002
    head:
      summary: Inspect durable ID or retained live handle (headers only)
      operationId: head_runs_id
      responses:
        '200':
          description: Response
        '404':
          description: Resource not found
        '500':
          description: Execution/inspection infrastructure error
        '405':
          description: Registered path with unsupported method; Allow header identifies
            supported methods
          headers: *id015
        '400':
          description: Invalid request
      parameters: *id016
...
components:
  schemas:
    Error:
      type: object
      additionalProperties: false
      required:
      - error
      properties:
        error: *id002
...
    RunHandle:
      type: object
      additionalProperties: false
      required:
      - run_id
      properties:
        run_id: *id002
    Cancellation:
      type: object
      additionalProperties: false
      required:
      - run_id
      - status
      properties:
        run_id: *id002
        status:
          const: cancelling
...
    PlanWarning:
      type: object
      additionalProperties: false
      required:
      - kind
      - node
      - param
      - message
      properties:
        kind: *id002
        node: *id002
        param: *id002
        message: *id002
...
    PartitionSummary:
      type: object
      additionalProperties: false
      required:
      - total
      - cached
      - will_run
      - will_run_keys
      properties:
        total: *id027
        cached: *id027
        will_run: *id027
        will_run_keys:
          type: array
          items: *id002
          maxItems: 20
    StepReport:
      type: object
      additionalProperties: false
      required:
      - id
      - kind
      properties:
        id: *id002
        kind:
          type: string
          enum:
          - asset
          - task
          - sensor
        action: *id002
        status: *id002
        reason: *id002
        detail: *id002
        run_hash: *id002
        artifact: *id002
        warning: *id002
        artifact_mismatch: *id026
        partitions:
          $ref: '#/components/schemas/PartitionSummary'
        env:
          type: object
          additionalProperties:
            anyOf:
            - *id002
            - type: 'null'
    ArtifactReference:
      type: object
      additionalProperties: false
      required:
      - path
      - format
      - size_bytes
      properties:
        path: *id002
        format: *id002
        size_bytes: *id027
        elapsed_seconds: &id028
          type: number
    Result:
      type: object
      additionalProperties: false
      required:
      - run_id
      - elapsed_seconds
      - steps_executed
      - phases
      - final_output
      - steps
      - warnings
      properties:
        run_id: *id002
        elapsed_seconds: *id028
        steps_executed: *id027
        phases: *id027
        final_output:
          anyOf:
          - $ref: '#/components/schemas/ArtifactReference'
          - type: 'null'
          description: Null or an artifact reference, not the CLI inline JSON rendering.
            This endpoint does not define a partition retrieval API.
        steps:
          type: array
          items:
            $ref: '#/components/schemas/StepReport'
        warnings:
          type: array
          items:
            $ref: '#/components/schemas/PlanWarning'
...
    Log:
      type: object
      additionalProperties: false
      required:
      - node_id
      - seq
      - line
      properties:
        node_id: *id002
        seq: *id037
        line: *id002
...
    RunSummary:
      type: object
      additionalProperties: false
      required:
      - id
      - handle
      - run_id
      - command
      - files
      - target
      - status
      - started_at
      - finished_at
      - elapsed_seconds
      - steps_total
      - steps_executed
      - steps_cached
      - error
      properties:
        id: *id002
        handle:
          anyOf:
          - *id002
          - type: 'null'
        run_id:
          anyOf:
          - *id002
          - type: 'null'
        command: *id002
        files:
          type: array
          items: *id002
        target:
          anyOf:
          - *id002
          - type: 'null'
        status:
          type: string
          description: Durable running/success/failed/cancelled/interrupted; pending
            or complete can appear for live runs without a persisted row. This differs
            from polling RunStatus.
        started_at: *id002
        finished_at:
          anyOf:
          - *id002
          - type: 'null'
        elapsed_seconds:
          anyOf:
          - *id028
          - type: 'null'
        steps_total:
          anyOf:
          - *id037
          - type: 'null'
        steps_executed: *id037
        steps_cached: *id037
        error:
          anyOf:
          - *id002
          - type: 'null'
...
    RunStep:
      type: object
      additionalProperties: false
      required:
      - node_id
      - status
      - elapsed_seconds
      - error
      properties:
        node_id: *id002
        status: *id002
        elapsed_seconds:
          anyOf:
          - *id028
          - type: 'null'
        error:
          anyOf:
          - *id002
          - type: 'null'
    RunDetail:
      type: object
      additionalProperties: false
      required:
      - run
      - steps
      - logs
      - result
      properties:
        run:
          $ref: '#/components/schemas/RunSummary'
        steps:
          type: array
          items:
            $ref: '#/components/schemas/RunStep'
        logs:
          type: array
          items:
            $ref: '#/components/schemas/Log'
        result:
          anyOf:
          - $ref: '#/components/schemas/Result'
          - type: 'null'
...